refactor: segment in feature oriented architecture
